+++ app/Http/Controllers/StateController.php
@@ -1,5 +1,4 @@
 <?php
-
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
@@ -7,7 +6,9 @@ use Illuminate\Http\Request;
 use App\Http\Requests;
 use App\Http\Controllers\Controller;
 use App\Customization\Eloquent\State;
+use App\Workflow\Eloquent\Definition;
 use App\Project\Provider;
+use DB;
 
 class StateController extends Controller
 {
@@ -19,6 +20,27 @@ class StateController extends Controller
     public function index($project_key)
     {
         $states = Provider::getStateList($project_key);
+        foreach ($states as $key => $state)
+        {
+            $workflows = Definition::whereRaw([ 'state_ids' => isset($state['key']) ? $state['key'] : $state['_id'] ])
+                             ->orderBy('project_key', 'asc')
+                             ->get([ 'project_key', 'name' ])
+                             ->toArray();
+            $states[$key]['workflows'] = $workflows;
+
+            if ($workflows)
+            {
+                $states[$key]['is_used'] = true;
+            }
+            else
+            {
+                $states[$key]['is_used'] = $this->isFieldUsedByIssue($project_key, 'state', $state); 
+            }
+
+            $states[$key]['workflows'] = array_filter($workflows, function($item) use($project_key) { 
+                                             return $item['project_key'] === $project_key || $item['project_key'] === '$_sys_$';
+                                         });
+        }
         return Response()->json(['ecode' => 0, 'data' => $states]);
     }
 
@@ -105,6 +127,18 @@ class StateController extends Controller
         if (!$state || $project_key != $state->project_key)
         {
             throw new \UnexpectedValueException('the state does not exist or is not in the project.', -10002);
+        }
+
+        $isUsed = $this->isFieldUsedByIssue($project_key, 'state', $state->toArray()); 
+        if ($isUsed)
+        {
+            throw new \UnexpectedValueException('the state has been used by issue.', -10002);
+        }
+
+        $isUsed = Definition::whereRaw([ 'state_ids' => isset($state->key) ? $state->key : $id ])->exists();
+        if ($isUsed)
+        {
+            throw new \UnexpectedValueException('the state has been used by workflow.', -10002);
         }
 
         State::destroy($id);
